Add crawllibrary function(using arrow notation to set "this" to this class). The content is currently just a placeholder.
The crawlLibrary function is supposed to send a GET request to the server to get it to crawl the library with the keyword and return the receive the response whichi contains the result of the search. The result of the search will be stored in the "crawl_result_array" state.